pyproject.toml
[tool.poetry]
name = "musicfa-bot"
version = "0.1.0"
description = ""
authors = ["Shervin <Shervinsaeedi78@gmail.com>"]
readme = "README.md"
packages = [{include = "musicfa_bot"}]

[tool.poetry.dependencies]
python = "3.10.7"
aiohttp = "^3.8.3"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
python-telegram-bot = {version = "^20.0a4", allow-prereleases = true}


[tool.poetry.group.dev.dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
